Background: Improve Old History Conversion
Make backup when converting old local history to new synced storage.
In case something goes south, we can instruct users to restore it.

If local history is found, it is converted to new palette in synced
storage. It is possible with multiple machines connected to same
synced storage.

If current palette is empty, user is switched to one with converted
colors.