Fix loading default palette if there is converted with colors
With 5.0 version I decided to switch to default whenever Plus is not activated
but forgot that there are people who are working in converted palette by default.

Their colors seams lost after upgrade to 5.0.

This will fix the behaviour for nonPlus users - default palette is 'converted'
now for those who have some colors in it and have empty default palette.

Fixes #106