Switch to contributions through Open Collective and donations change
Because Chrome Web Store Payments deprecation (see #164) I decided to try to use
Open Collective. It seems much better for my use case as I can focus on coding
and not accounting :)

And with that change I removed coin tipper widget, Paypal buttons and with it
finally long postponed jQuery.