fix: use onInstalled listener sooner
Because previous fix to ensure that opt-out is working, onInstalled is
now getting called too late in process. That results in situation, where
it isn't called in most situation.

I now switched to simply flagging extension as updated when UPDATE
reason is fired and use this flag later, when settings are already
loaded.