fix: do not open same campaign when already opened fixes #248
In the end, problem wasn't in function which verified that campaign has
already been opened, but in stored data.

I converted history format to simple array, so campaign ids are not
doubled. I also moved the code to own functions.

With tests added in previous commits, I hope everything should be OK.